5/28/24
Did some small changes to make the code a little less cluttered.
Currently working on the combat system, first step was to detect from which direction the player is clicking. 
Am able to detect that but currently trying to change it so that it only registers one direction that's the furthest.